Use of implementations in Services
